requirements.txt
fastapi
uvicorn
pymilvus
sentence-transformers
requests
beautifulsoup4
pillow
pytesseract
pdfplumber
groq
python-dotenv
streamlit
google-generativeai